+++ ColetaBlu/ColetaBlu/DTO/CadastroPontoDTO.cs
@@ -1,8 +1,7 @@
-﻿namespace ColetaBlu.Entity
+﻿namespace ColetaBlu.DTO
 {
-    public class PointRegistrationEntity
+    public class CadastroPontoDTO
     {
-        public int? Id { get; set; }
         public string? Name { get; set; }
         public string? Street { get; set; }
         public int? Number { get; set; }
