+++ ColetaBlu/ColetaBlu/Infrastructure/Connection.cs
@@ -5,9 +5,8 @@ namespace ColetaBlu.Infrastructure
 {
     public class Connection
     {
-        protected string connectionString = "Server=localhost;DataBase=coleta_blu;User=root;Password=root";
-
-
+        protected string connectionString = "Server=localhost;DataBase=coleta_blu;User=root;Password=@Thiago01";
+       
         protected MySqlConnection GetConnection()
         {
          return new MySqlConnection(connectionString);
