+++ site/Esquecidos/load.js
@@ -21,5 +21,3 @@ $('body').on('click', '.closeModal', (event) => {
 
 
 
-
-
