+++ site/css/meuspontos.css
@@ -1,171 +1,27 @@
-@font-face {
-
-    font-family: "mukta_bold";
-    src: url("../assets/fonts/pontoColeta/Mukta-Bold.ttf")
-}
-
-
-@font-face {
-
-    font-family: "mukta_medium";
-    src: url("../assets/fonts/pontoColeta/Mukta-Medium.ttf")
-}
-
-body {
-    background-color: #FAF3E1;
-}
-.main{
-    margin: auto;
-}
-.contanerMeusPontos h1 {
-    font-family: mukta_bold;
-    font-size: 20px;
-}
-
-.contanerMeusPontos h2 {
-    font-family: mukta_bold;
-    font-size: 15px;
-}
-
-.contanerMeusPontos h3 {
-    font-family: mukta_bold;
-    font-size: 15px;
-}
-
-.contanerMeusPontos p {
-    font-family: mukta_medium;
-    font-size: 15px;
-    color: #000;
-}
-
-.contanerMeusPontos {
-    color: #6A994E;
-    
-}
 .sectionMeusPontos{
     display: flex;
-    justify-content: center;
 }
-.sectionMeusPontos img{
-    display: none;
-}
-.addMeuPonto {
+.addMeuPonto{
+
     display: flex;
     justify-content: center;
 }
 
-.btnAddMeuPonto {
+.btnAddMeuPonto{
     width: 50px;
     height: 50px;
     border: 0px;
     border-radius: 25px;
     background-color: #6A994E;
-    margin: 50px;
 }
-
-.btnAddMeuPonto img {
-    position: relative;
-    width: 40px;
-    padding-top: 3px;
-    padding-right: 3px;
-    filter: brightness(0) invert(1) grayscale(100%);
-}
-.cardMeuPonto{
-    width: 350px;
-    height: 450px;
-    background-image: url(../assets/img/estabelecimento.jpg);
-    background-size: 350px 200px;
-    padding-top: 200px;
-    background-repeat: no-repeat;
-    border-top: 20px solid;
-    border-radius: 20px;
-    margin: auto;
-    box-shadow: 0px 4px 4px 0px rgba(0, 0, 0, 0.25) inset, 0px 4px 4px 0px rgba(0, 0, 0, 0.25);
-}
-
-
-.infoLocate{
-    width: 100%;
-    margin-left: 15px;
-}
-.infoResidue{
-    width: 100%;
-    margin: auto;
-}
-
-.infoResiduoColeta{
-    display: flex;
-    justify-content: center;
-    padding-left: 10px;
-}
-.infoResidue h2{
-    text-align: center;
-    padding-right: 20px;
-}
-.cardResidue {
+.btnAddMeuPonto:hover{
     width: 150px;
-    height: 150px;
-    margin: 10px;
-    border-radius: 20px;
-    background-color: #ECEAEA;
-    
+    transition: 0.5s;
 }
-@media (max-width: 1200px){
-    .infoResidue{
-        width: 200px;
-    }
-
+.textaddMeuPonto{
+    opacity: 0;
 }
-@media(min-width: 940px){
-    .contanerMeusPontos h1 {
-        font-family: mukta_bold;
-        font-size: 40px;
-    }
+.textaddMeuPonto:hover{
+    opacity: 1;
     
-    .contanerMeusPontos h2 {
-        font-family: mukta_bold;
-        font-size: 30px;
-    }
-    
-    .contanerMeusPontos h3 {
-        font-family: mukta_bold;
-        font-size: 20px;
-    }
-    
-    .contanerMeusPontos p {
-        font-family: mukta_medium;
-        font-size: 20px;
-        color: #000;
-    }
-    .cardMeuPonto {
-        background-image: url(../assets/img/estabelecimento.jpg);
-        background-repeat: no-repeat;
-        background-size: 300px 350px;
-        padding-left: 300px;
-        border-left: 20px solid;
-        padding-top: 0px;
-        border-top: 0px;
-        width: 60%;
-        max-width: 800px;
-        height: 350px;
-        columns:  2;
-        column-gap: 15px;
-        background-color: #FAF3E1;
-        display: flex;
-        margin: auto;
-        border-radius: 20px;
-        box-shadow: 0px 4px 4px 0px rgba(0, 0, 0, 0.25) inset, 0px 4px 4px 0px rgba(0, 0, 0, 0.25);
-    }
-    .infoLocate{
-        width: 50%;
-    }
-    .infoResidue{
-        width: 50%;
-        margin: 10px;
-    }
-    .infoResidue h2{
-        text-align: center;
-        padding-top: 15px;
-    }
-
 }
