+++ site/css/pontoColeta.css
@@ -12,7 +12,7 @@
 }
 
 .main {
-   margin: auto;
+    margin-top: 50px;
 }
 
 #coletaBluImg {
@@ -294,12 +294,12 @@
 }
 
 @media(min-width: 750px) {
-    
 
-   .mains{
-    width: 1500px;
-    margin: auto;
-   }
+    .main {
+        margin-left: 250px;
+
+
+    }
 
     .modalContainerPontoColeta.ativo {
         display: block;
