+++ site/index.html
@@ -34,14 +34,12 @@
     <div id="modalLogin" class="">
       <div id="backgroundLogin" class="flex">
         <div>
-
           <button class="closeModal" onclick="CLOSE_MODAL_LOGIN()">
             <div class="menuLogin">
               <span class="barLogin"></span>
               <span class="barLogin"></span>
             </div>
           </button>
-
           <div class="flex">
             <img id="logo" class="h-20" src="./assets/img/MicrosoftTeams-image (7).png" alt="Logo da ColetaBlu">
             <h1 class=" mt-10 enterLogin">Entrar</h1>
