+++ site/js/login.js
@@ -1,4 +1,3 @@
-
 function SHOW_MODAL_LOGIN() {
     let modal = document.querySelector('#modalLogin')
     modal.classList.add(`modalLoginActive`);
@@ -10,7 +9,6 @@ function CLOSE_MODAL_LOGIN() {
 }
 
 // teste
-
 window.addEventListener("load", (event) => {
 
     const usuario = document.querySelector(`#insertUserLogin`)
@@ -75,8 +73,4 @@ window.addEventListener("load", (event) => {
         }
 
     })
-
 })
-
-
-
