+++ site/meusPontos.html
@@ -1,58 +1,30 @@
 <!DOCTYPE html>
-<html lang="pt-br">
+<html lang="en">
+
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link rel="stylesheet" href="./css/meuspontos.css">
-    <link rel="stylesheet" href="./lib/owlCarousel/dist/assets/owl.carousel.min.css">
-    <link rel="stylesheet" href="./lib/owlCarousel/dist/assets/owl.theme.default.min.css">
     <title>Document</title>
-
-
 </head>
 
 <body>
-    <main class="contanerMeusPontos">
+    <main>
         <section class="sectionMeusPontos">
-            <div class="sectionMeusPontosText">
+            <div>
                 <h1>Meus estabelecimentos</h1>
                 <p>Veja e edite todos os seus estabelecimentos que realizam coletas de resíduos.</p>
             </div>
-
-            <img src="./assets/img/circulo-de-residuos 1.png" alt="">
-
+            <div>
+                <img src="./assets/img/circulo-de-residuos 1.png" alt="">
+            </div>
         </section>
         <div class="addMeuPonto">
-            <button class="btnAddMeuPonto"><a href="cadastroPonto.html" ><img src="./assets/img/mais.png" alt="" ></a></button>
+            <button class="btnAddMeuPonto"><span class="textaddMeuPonto">Novo ponto de coleta</span ></button>
         </div>
-        <div class="cardMeuPonto">
-            <div class="infoLocate">
-                <!-- <h2>Nome do Estabelecimento</h2>
-                <h3>Endereço</h3>
-                <p>rua</p>
-                <h3>Observação</h3>
-                <p>.............</p> -->
-            </div> 
-            <div class="infoResidue">
-                <h2>É coletado</h2> <br>
-                <section class="owl-carousel infoResiduoColeta">
-                    <!-- <div class="item cardResidue">
-                        <h3>Patrocinador X</h3>
-                        <lu>
-                            <li></li>
-                            <li></li>
-                            <li></li>
-                            <li></li>
-                            <li></li>
-                        </lu>
-                    </div> -->
-                </section>
-            </div>
-        </div>
+
 
     </main>
 </body>
-<script src="./lib/jquery/jquery.min.js"></script>
-<script src="./lib/owlCarousel/dist/owl.carousel.min.js"></script>
-<script src="./js/meusPontos.js"></script>
+
 </html>
